-refactor AppointmentDeletedEvent and NewAppointmentAddedEvent
