+++ src/main/java/seedu/address/commons/events/model/NewAppointmentAddedEvent.java
@@ -1,4 +1,4 @@
-package seedu.address.commons.events.ui;
+package seedu.address.commons.events.model;
 
 import seedu.address.commons.events.BaseEvent;
 import seedu.address.model.appointment.Appointment;
